m16_06_02/pyproject.toml
[tool.poetry]
name = "m16-06-02"
version = "0.1.0"
description = ""
authors = ["Yuriy Kuchma <krabatua@gmail.com>"]
readme = "README.md"
packages = [{include = "m16_06_02"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
faker = "^19.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
